+++ packages/material-ui-codemod/README.md
@@ -97,7 +97,7 @@ The list includes these transformers
 - [`modal-props`](#modal-props)
 - [`moved-lab-modules`](#moved-lab-modules)
 - [`pagination-round-circular`](#pagination-round-circular)
-- [`private-imports`](#private-imports)
+- [`optimal-imports`](#optimal-imports)
 - [`root-ref`](#root-ref)
 - [`skeleton-variant`](#skeleton-variant)
 - [`styled-engine-provider`](#styled-engine-provider)
@@ -719,7 +719,7 @@ npx @material-ui/codemod@next v5.0.0/pagination-round-circular <path>
 
 You can find more details about this breaking change in [the migration guide](https://next.material-ui.com/guides/migration-v4/#pagination).
 
-#### `private-imports`
+#### `optimal-imports`
 
 Fix private import paths.
 
@@ -733,7 +733,7 @@ Fix private import paths.
 <!-- #default-branch-switch -->
 
 ```sh
-npx @material-ui/codemod@next v5.0.0/private-imports <path>
+npx @material-ui/codemod@next v5.0.0/optimal-imports <path>
 ```
 
 #### `root-ref`
