+++ packages/material-ui-codemod/src/v5.0.0/optimal-imports.test.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import { expect } from 'chai';
 import jscodeshift from 'jscodeshift';
-import transform from './private-imports';
+import transform from './optimal-imports';
 import readFile from '../util/readFile';
 
 function trim(str) {
@@ -15,21 +15,21 @@ function read(fileName) {
 
 describe('@material-ui/codemod', () => {
   describe('v5.0.0', () => {
-    describe('private-imports', () => {
+    describe('optimal-imports', () => {
       it('convert path as needed', () => {
         const actual = transform(
-          { source: read('./private-imports.test/actual.js'), path: require.resolve('./private-imports.test/actual.js') },
+          { source: read('./optimal-imports.test/actual.js'), path: require.resolve('./optimal-imports.test/actual.js') },
           { jscodeshift: jscodeshift },
           {},
         );
 
-        const expected = read('./private-imports.test/expected.js');
+        const expected = read('./optimal-imports.test/expected.js');
         expect(trim(actual)).to.equal(trim(expected), 'The transformed version should be correct');
       });
 
       it('should be idempotent', () => {
         const actual = transform(
-          { source: read('./private-imports.test/expected.js'), path: require.resolve('./private-imports.test/expected.js') },
+          { source: read('./optimal-imports.test/expected.js'), path: require.resolve('./optimal-imports.test/expected.js') },
           { jscodeshift: jscodeshift },
           {},
         );
